Move app_controller.py into RimSort.controllers
